refactor(desktop): extract logger init to dedicated logger.ts
- Move all electron-log configuration (file rotation, sanitization,
  truncation, rate limiting) from app.ts to logger.ts
- Simplify renderer index.desktop.ts to just forward raw messages;
  all filtering now runs in main process via logger.hooks and format
- app.ts reduced by ~150 lines, index.desktop.ts by ~100 lines
