+++ apps/desktop/app/app.ts
@@ -65,155 +65,8 @@ import { initSentry } from './sentry';
 import { startServices } from './service';
 import { setMainWindowForOAuthServer } from './service/oauthLocalServer/oauthLocalServer';
 
-logger.initialize();
-
-// Align file logging with react-native-native-logger behavior:
-// - File name: app-latest.log (matches iOS/Android)
-// - Max size: 20MB per file (matches native MAX_FILE_SIZE)
-// - Rolling: date-based archives as app-{date}.{i}.log
-// - Max history: 6 archived files (matches native MAX_HISTORY)
-logger.transports.file.fileName = 'app-latest.log';
-logger.transports.file.maxSize = 20 * 1024 * 1024;
-
-// Custom archive function: rename rolled log to app-{yyyy-MM-dd}.{i}.log
-// and clean up archives beyond MAX_HISTORY, matching native behavior.
-const MAX_LOG_HISTORY = 6;
-logger.transports.file.archiveLogFn = (oldLogFile: {
-  path: string;
-  toString(): string;
-}) => {
-  const oldPath = oldLogFile.path || oldLogFile.toString();
-  const dir = path.dirname(oldPath);
-  const now = new Date();
-  const pad2 = (n: number) => String(n).padStart(2, '0');
-  const dateStr = `${now.getFullYear()}-${pad2(now.getMonth() + 1)}-${pad2(now.getDate())}`;
-
-  // Find next available index using link+unlink for cross-platform safety.
-  // On Linux, renameSync silently overwrites the target, so we use
-  // linkSync (which throws EEXIST atomically) to avoid overwriting archives.
-  let index = 0;
-  let moved = false;
-  while (!moved && index < 1000) {
-    const archivePath = path.join(dir, `app-${dateStr}.${index}.log`);
-    try {
-      fs.linkSync(oldPath, archivePath);
-      fs.unlinkSync(oldPath);
-      moved = true;
-    } catch (e: any) {
-      // eslint-disable-next-line @typescript-eslint/no-unsafe-member-access
-      if (e?.code === 'EEXIST') {
-        // Collision: archive slot taken; try next index
-        index += 1;
-        continue;
-      }
-      // linkSync not supported (e.g. cross-device) — fall back to renameSync
-      // eslint-disable-next-line @typescript-eslint/no-unsafe-member-access
-      if (e?.code === 'EXDEV') {
-        try {
-          fs.renameSync(oldPath, archivePath);
-          moved = true;
-        } catch {
-          // eslint-disable-next-line no-console
-          console.warn('[log-archive] Failed to archive log file:', e);
-          break;
-        }
-      } else {
-        // eslint-disable-next-line no-console
-        console.warn('[log-archive] Failed to archive log file:', e);
-        break;
-      }
-    }
-    index += 1;
-  }
-  if (!moved) {
-    // eslint-disable-next-line no-console
-    console.warn('[log-archive] Could not archive after 1000 attempts');
-  }
-
-  // Cleanup: remove oldest archives beyond MAX_HISTORY
-  try {
-    const files = fs
-      .readdirSync(dir)
-      .filter(
-        (f: string) =>
-          f.startsWith('app-') && f.endsWith('.log') && f !== 'app-latest.log',
-      )
-      .sort()
-      .reverse();
-
-    if (files.length > MAX_LOG_HISTORY) {
-      for (const file of files.slice(MAX_LOG_HISTORY)) {
-        try {
-          fs.unlinkSync(path.join(dir, file));
-        } catch {
-          // ignore cleanup errors
-        }
-      }
-    }
-
-    // Clean up legacy main.log / main.old.log files from previous electron-log defaults
-    for (const legacyFile of ['main.log', 'main.old.log']) {
-      try {
-        fs.unlinkSync(path.join(dir, legacyFile));
-      } catch {
-        // ignore if not exists
-      }
-    }
-  } catch {
-    // ignore cleanup errors
-  }
-};
-
-// Sanitize sensitive data in main process logs, matching native NativeLogger.sanitize()
-const MAIN_PROCESS_SENSITIVE_PATTERNS = [
-  /(?:0x)?[0-9a-fA-F]{64}/g,
-  /\b[5KL][1-9A-HJ-NP-Za-km-z]{50,51}\b/g,
-  /\b[xyzXYZ](?:prv|pub)[1-9A-HJ-NP-Za-km-z]{107,108}\b/g,
-  /(?:\b[a-z]{3,8}\b[\s,]+){11,}\b[a-z]{3,8}\b/g,
-  /(?:Bearer|token[=:]?)\s*[A-Za-z0-9_.\-+/=]{20,}/g,
-  /(?:eyJ|AAAA)[A-Za-z0-9+/=]{40,}/g,
-];
-
-function sanitizeMainProcess(message: string): string {
-  let result = message;
-  for (const pattern of MAIN_PROCESS_SENSITIVE_PATTERNS) {
-    pattern.lastIndex = 0;
-    result = result.replace(pattern, '[REDACTED]');
-  }
-  return result.replace(/\n/g, ' ').replace(/\r/g, ' ');
-}
-
-function sanitizeDataArray(data: any[]): any[] {
-  return data.map((item) => {
-    if (typeof item === 'string') {
-      return sanitizeMainProcess(item);
-    }
-    return item;
-  });
-}
-
-// Custom file format: app-scoped messages from the renderer (react-native-logs)
-// already contain timestamps and level info, so write them as-is to match
-// the output format of react-native-native-logger on mobile.
-// Main process messages keep the standard electron-log prefix and are sanitized.
-// eslint-disable-next-line @typescript-eslint/no-unsafe-return
-logger.transports.file.format = (params: {
-  data: any[];
-  level: string;
-  message: { date: Date; scope?: string };
-}) => {
-  if (params.message?.scope === 'app') {
-    // App-scoped messages are already sanitized in the renderer
-    // eslint-disable-next-line @typescript-eslint/no-unsafe-return
-    return params.data;
-  }
-  const d = params.message.date;
-  const pad2 = (n: number) => String(n).padStart(2, '0');
-  const pad3 = (n: number) => String(n).padStart(3, '0');
-  const ts = `${d.getFullYear()}-${pad2(d.getMonth() + 1)}-${pad2(d.getDate())} ${pad2(d.getHours())}:${pad2(d.getMinutes())}:${pad2(d.getSeconds())}.${pad3(d.getMilliseconds())}`;
-  // eslint-disable-next-line @typescript-eslint/no-unsafe-return
-  return [`[${ts}] [${params.level}]`, ...sanitizeDataArray(params.data)];
-};
+// Logger initialization (file rotation, sanitization, rate limiting)
+import './logger';
 
 initSentry();
